Move our toolbar patternfly override to namespace-list, only used there now
Issue: AAH-820
