+++ CHANGES/820.misc
@@ -1,0 +1,1 @@
+Namespaces list: use same filter as other screens, remove custom Toolbar override
