+++ src/containers/namespace-list/namespace-list.tsx
@@ -1,11 +1,21 @@
-import { t } from '@lingui/macro';
 import * as React from 'react';
-import './namespace-list.scss';
-
+import {
+  Button,
+  ButtonVariant,
+  InputGroup,
+  TextInput,
+  Toolbar,
+  ToolbarContent,
+  ToolbarGroup,
+  ToolbarItem,
+} from '@patternfly/react-core';
 import { RouteComponentProps, Redirect } from 'react-router-dom';
+import { SearchIcon } from '@patternfly/react-icons';
+import { t } from '@lingui/macro';
 
 import { ParamHelper } from 'src/utilities/param-helper';
 import {
+  AlertList,
   BaseHeader,
   EmptyStateFilter,
   EmptyStateNoData,
@@ -15,16 +25,17 @@ import {
   NamespaceCard,
   NamespaceModal,
   Pagination,
-  Toolbar,
-  AlertList,
+  Sort,
+  SortFieldType,
 } from 'src/components';
-import { Button, ToolbarItem } from '@patternfly/react-core';
 import { NamespaceAPI, NamespaceListType, MyNamespaceAPI } from 'src/api';
 import { formatPath, namespaceBreadcrumb, Paths } from 'src/paths';
 import { Constants } from 'src/constants';
 import { filterIsSet } from 'src/utilities';
 import { AppContext } from 'src/loaders/app-context';
 import { i18n } from '@lingui/core';
+
+import './namespace-list.scss';
 
 interface IState {
   namespaces: NamespaceListType[];
@@ -197,7 +208,7 @@ export class NamespaceList extends React.Component<IProps, IState> {
           )}
           {noData ? null : (
             <div className='toolbar'>
-              <Toolbar
+              <NamespaceListToolbar
                 params={params}
                 sortOptions={[{ title: t`Name`, id: 'name', type: 'alpha' }]}
                 searchPlaceholder={search}
@@ -340,3 +351,108 @@ export class NamespaceList extends React.Component<IProps, IState> {
 }
 
 NamespaceList.contextType = AppContext;
+
+interface IxProps {
+  /** Current page params */
+  params: {
+    sort?: string;
+    keywords?: string;
+  };
+
+  /** List of sort options that the user can pick from */
+  sortOptions?: SortFieldType[];
+
+  /** Sets the current page params to p */
+  updateParams: (params) => void;
+
+  /** Search bar placeholder text*/
+  searchPlaceholder: string;
+
+  /** Extra set of customizeable inputs that appear to right of sort*/
+  extraInputs?: React.ReactNode[];
+}
+
+interface IxState {
+  kwField: string;
+}
+
+export class NamespaceListToolbar extends React.Component<IxProps, IxState> {
+  static defaultProps = {
+    extraInputs: [],
+  };
+
+  constructor(props) {
+    super(props);
+    this.state = {
+      kwField: props.params.keywords || '',
+    };
+  }
+
+  render() {
+    const {
+      params,
+      sortOptions,
+      updateParams,
+      searchPlaceholder,
+      extraInputs,
+    } = this.props;
+    const { kwField } = this.state;
+    return (
+      <Toolbar>
+        <ToolbarContent>
+          <ToolbarGroup style={{ marginLeft: 0 }}>
+            <ToolbarItem>
+              <InputGroup>
+                <TextInput
+                  value={kwField}
+                  onChange={(k) => this.setState({ kwField: k })}
+                  onKeyPress={(e) => this.handleEnter(e)}
+                  type='search'
+                  aria-label={t`search text input`}
+                  placeholder={searchPlaceholder}
+                />
+                <Button
+                  variant={ButtonVariant.control}
+                  aria-label={t`search button`}
+                  onClick={() => this.submitKeywords()}
+                >
+                  <SearchIcon />
+                </Button>
+              </InputGroup>
+            </ToolbarItem>
+          </ToolbarGroup>
+          {sortOptions && (
+            <ToolbarGroup>
+              <ToolbarItem>
+                <Sort
+                  options={sortOptions}
+                  params={params}
+                  updateParams={updateParams}
+                />
+              </ToolbarItem>
+            </ToolbarGroup>
+          )}
+          {extraInputs}
+        </ToolbarContent>
+      </Toolbar>
+    );
+  }
+
+  private handleEnter(e) {
+    // l10n: don't translate
+    if (e.key === 'Enter') {
+      this.submitKeywords();
+    }
+  }
+
+  private submitKeywords() {
+    this.props.updateParams({
+      ...ParamHelper.setParam(
+        this.props.params,
+        'keywords',
+        this.state.kwField,
+      ),
+      page: 1, // always reset the page when searching
+    });
+  }
+}
